Modificados archivos para que muestren el id público y modificado el createRestaurant para que cree un usuario administrador asociado

+++ backend/app/Family/Application/DeleteFamilyByID/DeleteFamilyByID.php
@@ -3,6 +3,7 @@
 namespace App\Family\Application\DeleteFamilyByID;
 
 use App\Family\Domain\Interfaces\FamilyRepositoryInterface;
+use InvalidArgumentException;
 
 class DeleteFamilyByID
 {
@@ -12,6 +13,11 @@ class DeleteFamilyByID
 
     public function __invoke(string $id): void
     {
-        $deleted = $this->familyRepository->deleteByID($id);
+        $exists = $this->familyRepository->findById($id);
+        if($exists === null) {
+            throw new InvalidArgumentException("Family doesn't exist or is already deleted.");
+        } else {
+            $deleted = $this->familyRepository->deleteByID($id);
+        }
     }
 }
